requirements.txt
flask==2.3.3
flask-cors==4.0.0
praw==7.7.1
requests==2.31.0
vaderSentiment==3.3.2
python-dotenv==1.0.1